Pod verb fixes
