+++ Content.Server/SS220/CultYogg/Pod/CultYoggPodSystem.cs
@@ -3,7 +3,6 @@
 using Content.Shared.DragDrop;
 using Content.Shared.SS220.CultYogg.Pod;
 using Content.Shared.SS220.CultYogg.MiGo;
-using Content.Shared.Verbs;
 using Robust.Shared.Containers;
 using Robust.Server.GameObjects;
 
@@ -20,41 +19,6 @@ public sealed partial class CultYoggPodSystem : SharedCultYoggPodSystem
         SubscribeLocalEvent<CultYoggPodComponent, EntRemovedFromContainerMessage>(GotRemoved);
         SubscribeLocalEvent<CultYoggPodComponent, EntityTerminatingEvent>(GotTerminated);
         SubscribeLocalEvent<CultYoggPodComponent, DragDropTargetEvent>(OnCanDropHandle);
-        SubscribeLocalEvent<CultYoggPodComponent, GetVerbsEvent<AlternativeVerb>>(AddInsertVerb);
-    }
-
-    private void AddInsertVerb(Entity<CultYoggPodComponent> ent, ref GetVerbsEvent<AlternativeVerb> args)
-    {
-        if (!args.CanAccess || !args.CanInteract)
-            return;
-
-        var target = args.User;
-
-        if (ent.Comp.MobContainer.ContainedEntity is not null)
-        {
-            AlternativeVerb verb = new()
-            {
-                Act = () => TryEject(ent.Comp.MobContainer.ContainedEntity.Value, ent),
-                Category = VerbCategory.Eject,
-                Text = Loc.GetString("cult-yogg-eject-pod"),
-                Priority = 1
-            };
-
-            args.Verbs.Add(verb);
-        }
-
-        if (ent.Comp.MobContainer.ContainedEntity is null)
-        {
-            AlternativeVerb verb = new()
-            {
-                Act = () => TryInsert(target, ent),
-                Category = VerbCategory.Insert,
-                Text = Loc.GetString("cult-yogg-ensert-pod"),
-                Priority = 1
-            };
-
-            args.Verbs.Add(verb);
-        }
     }
 
     private void OnCanDropHandle(Entity<CultYoggPodComponent> ent, ref DragDropTargetEvent args)
