+++ Content.Shared/SS220/CultYogg/Pod/SharedCultYoggPodSystem.cs
@@ -5,6 +5,9 @@ using Content.Shared.DragDrop;
 using Content.Shared.Mobs.Components;
 using Content.Shared.Popups;
 using Content.Shared.SS220.CultYogg.Cultists;
+using Content.Shared.SS220.CultYogg.MiGo;
+using Content.Shared.Verbs;
+using Content.Shared.Whitelist;
 using Robust.Shared.Containers;
 using Robust.Shared.Network;
 
@@ -15,12 +18,15 @@ public abstract class SharedCultYoggPodSystem : EntitySystem
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
     [Dependency] private readonly SharedContainerSystem _container = default!;
+    [Dependency] private readonly EntityWhitelistSystem _entityWhitelist = default!;
+
     public override void Initialize()
     {
         base.Initialize();
 
         SubscribeLocalEvent<CultYoggPodComponent, ComponentInit>(OnCompInit);
         SubscribeLocalEvent<CultYoggPodComponent, CanDropTargetEvent>(OnPodCanDrop);
+        SubscribeLocalEvent<CultYoggPodComponent, GetVerbsEvent<AlternativeVerb>>(AddInsertVerb);
     }
 
     private void OnPodCanDrop(Entity<CultYoggPodComponent> ent, ref CanDropTargetEvent args)
@@ -37,6 +43,38 @@ public abstract class SharedCultYoggPodSystem : EntitySystem
         ent.Comp.MobContainer = _container.EnsureContainer<ContainerSlot>(ent, "cultyYoggPod");
     }
 
+    private void AddInsertVerb(Entity<CultYoggPodComponent> ent, ref GetVerbsEvent<AlternativeVerb> args)
+    {
+        if (!args.CanAccess || !args.CanInteract)
+            return;
+
+        var target = args.User;
+
+        if (ent.Comp.MobContainer.ContainedEntity is not null)
+        {
+            AlternativeVerb verb = new()
+            {
+                Act = () => TryEject(ent.Comp.MobContainer.ContainedEntity.Value, ent),
+                Text = Loc.GetString("cult-yogg-eject-pod"),
+                Priority = 1
+            };
+
+            args.Verbs.Add(verb);
+        }
+
+        if (ent.Comp.MobContainer.ContainedEntity is null)
+        {
+            AlternativeVerb verb = new()
+            {
+                Act = () => TryInsert(target, ent),
+                Text = Loc.GetString("cult-yogg-ensert-pod"),
+                Priority = 1
+            };
+
+            args.Verbs.Add(verb);
+        }
+    }
+
     public bool TryInsert(EntityUid entToEnsert, Entity<CultYoggPodComponent> podEnt)
     {
         if (podEnt.Comp.MobContainer.ContainedEntity != null)
@@ -45,7 +83,7 @@ public abstract class SharedCultYoggPodSystem : EntitySystem
         if (!HasComp<MobStateComponent>(entToEnsert) || !HasComp<DamageableComponent>(entToEnsert))
             return false;
 
-        if (!HasComp<CultYoggComponent>(entToEnsert))
+        if (_entityWhitelist.IsWhitelistFail(podEnt.Comp.CultistsWhitelist, entToEnsert))
         {
             if (_net.IsServer)
                 _popup.PopupEntity(Loc.GetString("cult-yogg-heal-only-cultists"), entToEnsert, entToEnsert);
